backend/requirements.txt
psycopg2-binary
bcrypt==3.1.2
flask
Flask-Bcrypt==0.7.1
SQLAlchemy
pyyaml
PyJWT==1.4.2